create tiktokcardcontent component
